requirements.txt
# ai-sales-assistant/requirements.txt
# Core Streamlit application and data handling
streamlit==1.33.0
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.7.1

# Google APIs and Authentication
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# Firebase Database
firebase-admin==6.4.2

# Gemini API
google-generativeai==0.6.0

# API Call Robustness
tenacity==8.3.0
